# api/openapi.yaml
openapi: 3.0.3
info:
  title: Runcriticon API
  version: 0.1.0
  description: API REST de Runcriticon
servers:
  - url: /api
    description: Mismo origen que la SPA (cookie de sesión first-party, sin CORS)
...
security:
  - cookieAuth: []
paths:
...
  /activacion:
    post:
      summary: Activar la cuenta fijando contraseña con el token de invitación
      description: |
        Activa una cuenta invitada fijando su contraseña con el token del email (un solo uso) y, al
        lograrlo, inicia sesión (auto-login, cookie httpOnly). Endpoint anónimo. Devuelve 400/404 si
        el token es inválido o caducado, y 409 si la cuenta ya está activa.
      operationId: activarCuenta
      tags: [activacion]
      security: []
      requestBody:
        $ref: '#/components/requestBodies/ActivationBody'
      responses:
        '200':
          $ref: '#/components/responses/CuentaActivada'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          $ref: '#/components/responses/ErrorConflict'
        '500':
          $ref: '#/components/responses/ErrorServerError'
    get:
      summary: Resolver los detalles de una invitación por su token
      description: |
        Devuelve el nombre del invitado, el club, el rol asignado y quién invita, para pintar la
        tarjeta de contexto de la pantalla de activación antes de enviar el formulario.
        Endpoint anónimo. 400 si el token no coincide con ninguna invitación o ha caducado, 409 si
        ya fue consumida, 404 en el caso raro de que el usuario referenciado no exista. A diferencia
        de magic link/reseteo, sí devuelve 429 con Retry-After al agotar el límite: el token ya es el
        secreto de 256 bits, no hay un email que enumerar detrás de este endpoint.
      operationId: consultarInvitacion
      tags: [activacion]
      security: []
      parameters:
        - $ref: '#/components/parameters/InvitationTokenQuery'
      responses:
        '200':
          $ref: '#/components/responses/InvitacionResuelta'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          $ref: '#/components/responses/ErrorConflict'
        '429':
          $ref: '#/components/responses/ErrorTooManyRequests'
        '500':
          $ref: '#/components/responses/ErrorServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
      description: |
        Sesión autenticada por cookie httpOnly first-party (Spring Session). El navegador la adjunta
        automáticamente; el cliente nunca la lee ni la envía a mano. Los endpoints de entrada anónima
        (login, activación, magic link, reseteo) no la requieren (security: []).
  headers:
    RetryAfter:
      description: Segundos recomendados antes de reintentar
      schema:
        type: integer
  parameters:
...
    InvitationTokenQuery:
      name: token
      in: query
      required: true
      description: Token de invitación recibido por email
      schema:
        $ref: '#/components/schemas/Token'
...
  requestBodies:
...
    ActivationBody:
      description: Token de invitación y contraseña elegida
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ActivationRequest'
...
  responses:
...
    CuentaActivada:
      description: Cuenta activada; sesión iniciada (cookie httpOnly)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ActivationResponse'
    InvitacionResuelta:
      description: Detalles de la invitación, válida y aún no consumida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DetalleInvitacion'
...
    ErrorBadRequest:
      description: Entrada inválida (campo requerido vacío, formato incorrecto)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_INPUT
            field: email
            message: El email no tiene un formato válido
...
    ErrorNotFound:
      description: Recurso no encontrado o no pertenece al club del principal
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Recurso no encontrado
    ErrorConflict:
      description: La operación choca con el estado actual del recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: Ya existe un usuario con ese email en el club
    ErrorTooManyRequests:
      description: |
        Límite de rate-limiting alcanzado. Cuerpo ErrorResponse con `code=RATE_LIMITED` y cabecera
        `Retry-After` (segundos). Los flujos anónimos de magic link y reseteo NO devuelven 429:
        mantienen su 202 neutro para no revelar que se alcanzó un límite.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RATE_LIMITED
            message: Has superado el límite de peticiones; inténtalo de nuevo más tarde
    ErrorServerError:
      description: |
        Error interno del servidor (red de seguridad de GlobalRestExceptionHandler). Cuerpo
        ErrorResponse con `code=INTERNAL_ERROR` y mensaje neutro: el detalle del fallo va al log,
        nunca al body. Puede devolverlo cualquier endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: Ha ocurrido un error inesperado
...
  schemas:
...
    Nombre:
      type: string
      minLength: 1
      maxLength: 200
      description: Nombre completo de la persona
      example: Ana García
    NombreClub:
      type: string
      minLength: 1
      maxLength: 200
      description: Nombre del club
      example: Club Atletismo Central
    Token:
      type: string
      minLength: 1
      maxLength: 512
      description: Token de un solo uso recibido por email (invitación, magic link o reseteo)
      example: 8Kq2mZ7pR4tV9wX1yB6nC3dF5gH0jL
    Password:
      type: string
      format: password
      minLength: 12
      maxLength: 128
      description: Contraseña nueva; política de contraseña (12–128 caracteres, sin datos personales, no reutilizada)
    ActivationRequest:
      type: object
      description: |
        Datos para activar la cuenta por invitación. `consentimiento`/`versionConsentimiento`
        (consentimiento explícito Art. 9.2.a RGPD) solo son obligatorios cuando el invitado es
        ALUMNO: es el único interesado de los
        datos de salud del art. 9 RGPD que captura `seguimiento.reporte_sesion`. El backend los
        ignora para ADMIN/ENTRENADOR. `versionConsentimiento` debe coincidir con la versión vigente
        del texto (400 `CONSENTIMIENTO_REQUERIDO` sin casilla marcada; 409
        `VERSION_CONSENTIMIENTO_OBSOLETA` con una versión que ya no es la vigente).
      required: [token, password]
      properties:
        token:
          $ref: '#/components/schemas/Token'
        password:
          $ref: '#/components/schemas/Password'
        consentimiento:
          type: boolean
          description: Casilla de consentimiento de datos de salud, no premarcada. Solo aplica a ALUMNO.
        versionConsentimiento:
          type: string
          description: Versión del texto de consentimiento mostrado al marcar la casilla.
    ActivationResponse:
      description: Sesión iniciada tras la activación (mismo shape que la sesión de login)
      allOf:
        - $ref: '#/components/schemas/SessionResponse'
    DetalleInvitacion:
      type: object
      description: |
        Datos de contexto de una invitación válida y aún no consumida, para pintar la tarjeta de la
        pantalla de activación antes de enviar el formulario.
      required: [nombre, club, rol]
      properties:
        nombre:
          $ref: '#/components/schemas/Nombre'
        club:
          $ref: '#/components/schemas/NombreClub'
        rol:
          type: string
          enum: [ADMIN, ENTRENADOR, ALUMNO]
          description: Rol asignado a la persona invitada
          example: ALUMNO
        invitadoPor:
          type: string
          maxLength: 200
          nullable: true
          description: |
            Nombre completo de quien emitió la invitación. Ausente en invitaciones emitidas antes
            de que este dato empezara a registrarse.
          example: Ana Pinares
...
    SessionResponse:
      type: object
      description: Sesión iniciada; nunca incluye datos sensibles ni el hash de la contraseña
      required: [userId, clubId, role]
      properties:
        userId:
          type: string
          format: uuid
          example: 018f9a2b-1c3d-7e4f-8a9b-0c1d2e3f4a5b
        clubId:
          type: string
          format: uuid
          example: 018f9a2b-9f8e-7d6c-b5a4-3210fedcba98
        role:
          type: string
          enum: [ADMIN, ENTRENADOR, ALUMNO]
          description: Rol del usuario
          example: ENTRENADOR
...
    ErrorResponse:
      type: object
      description: |
        Error estructurado para respuestas 4xx y 5xx.
        El frontend traduce `code` a mensaje localizado; `field` solo aparece en errores de entrada.
      required: [code, message]
      properties:
        code:
          type: string
          description: Código estable para que el frontend lo traduzca
          example: INVALID_INPUT
        field:
          type: string
          nullable: true
          description: |
            Campo que originó el error, cuando hay uno concreto. Habitual en los 400 de validación,
            pero no exclusivo: `DUPLICATE_LABEL` lo lleva en un 409 para señalar qué campo choca.
          example: email
        message:
          type: string
          description: Mensaje legible (orientativo, no usar para lógica de cliente)
          example: El email no tiene un formato válido